-addressing comments on PR (moving interfaces to types) -removing interfaces from styles for now
